Refreshing

+++ root/usr/local/bin/avr.sh
@@ -5,11 +5,14 @@ sleep 2;
 FILES=`ls $DIR`
 for f in $FILES
 do
-	killall avrcontroller
-	echo "Processing $DIR/$f" > /tmp/avr
-	avrdude -c linuxspi -p m328 -P /dev/spidev0.0 -F -U flash:w:$DIR/$f 2>&1 /tmp/avr
-	rm -f $DIR/$f
+	RET=`lsof | grep $DIR/$f | wc -l`
+	if [ "$RET" -eq "0" ]; then
+		killall avrspi
+		echo "Processing $DIR/$f" > /tmp/avr
+		avrdude -c linuxspi -p m328 -P /dev/spidev0.0 -F -U flash:w:$DIR/$f 2>&1 /tmp/avr
+		rm -f $DIR/$f
+		/usr/local/bin/avrspi -d 
+	fi
 done
 
 done
-
